refactor imports, output_schema type check
